Added camera

+++ Valour.Engine/src/GpuBuffer.cs
@@ -8,13 +8,13 @@ internal class GpuBuffer<T> : IDisposable where T : unmanaged
     private bool _disposed;
 
     public uint Handle { get; }
-    public int Count { get; }
+    public uint Count { get; }
 
     public unsafe GpuBuffer(GL gl, ReadOnlySpan<T> data)
     {
         _gl = gl;
         
-        Count = data.Length;
+        Count = (uint)data.Length;
         
         _gl.CreateBuffers(1, out uint handle);
         Handle = handle;
